requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy>=2.0.25
alembic==1.12.1
psycopg2-binary>=2.9.0
redis==5.0.1
pydantic>=2.0.0
pydantic[email]>=2.0.0
pydantic-settings>=2.0.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
cryptography==41.0.7
PyJWT==2.8.0
bcrypt==4.1.2
httpx==0.25.2
aiofiles==23.2.0
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
hypothesis==6.92.1
python-dotenv==1.0.0
prometheus-client==0.19.0